separate missing sections logic into its own component. do some additional cleanup
